refactor auth code
