+++ management_api_app/resources/strings.py
@@ -21,12 +21,11 @@ STATE_STORE_ENDPOINT_NOT_RESPONDING = "State Store endpoint is not responding"
 UNSPECIFIED_ERROR = "Unspecified error"
 
 # Error strings
-AUTH_NOT_ASSIGNED_TO_ROLE = "Not assigned to any role"
+AUTH_NOT_ASSIGNED_TO_ADMIN_ROLE = "Not assigned to admin role"
 AUTH_COULD_NOT_VALIDATE_CREDENTIALS = "Could not validate credentials"
 WORKSPACE_DOES_NOT_EXIST = "Workspace does not exist"
 WORKSPACE_TEMPLATE_DOES_NOT_EXIST = "Workspace template does not exist"
 WORKSPACE_TEMPLATE_VERSION_EXISTS = "A template with this version already exists"
-
 
 # Resource Status
 RESOURCE_STATUS_NOT_DEPLOYED = "not_deployed"
