+++ management_api_app/services/aad_authentication.py
@@ -1,40 +1,23 @@
 import base64
 import logging
+from typing import List
 
 import jwt
 import requests
 import rsa
-
 from fastapi import Request, HTTPException, status
 from fastapi.security import OAuth2AuthorizationCodeBearer
+from pydantic import BaseModel, Field, parse_obj_as
 
 from core import config
 from resources import strings
 
 
-class User:
-    def __init__(self, decoded_token: dict):
-        self._claims = decoded_token
-
-    @property
-    def id(self):
-        return self._get_claim('oid')
-
-    @property
-    def name(self):
-        return self._get_claim('name')
-
-    @property
-    def email(self):
-        return self._get_claim('email')
-
-    @property
-    def roles(self):
-        return self._get_claim('roles')
-
-    def _get_claim(self, claim: str):
-        if claim in self._claims:
-            return self._claims[claim]
+class User(BaseModel):
+    oid: str
+    name: str
+    email: str
+    roles: List[str] = Field([])
 
 
 class AzureADAuthorization(OAuth2AuthorizationCodeBearer):
@@ -54,11 +37,7 @@ class AzureADAuthorization(OAuth2AuthorizationCodeBearer):
 
         try:
             decoded_token = self._decode_token(token)
-
-            if 'roles' not in decoded_token or decoded_token['roles'] == {}:
-                raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=strings.AUTH_NOT_ASSIGNED_TO_ROLE)
-
-            return User(decoded_token)
+            return parse_obj_as(User, decoded_token)
         except Exception as e:
             logging.debug(e)
             raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail=strings.AUTH_COULD_NOT_VALIDATE_CREDENTIALS, headers={"WWW-Authenticate": "Bearer"})
@@ -86,10 +65,8 @@ class AzureADAuthorization(OAuth2AuthorizationCodeBearer):
     def _get_token_key(self, key_id: str) -> str:
         if key_id not in self._jwt_keys:
             response = requests.get(f"{config.AAD_INSTANCE}/{config.TENANT_ID}/v2.0/.well-known/openid-configuration")
-
             aad_metadata = response.json() if response.ok else None
             jwks_uri = aad_metadata['jwks_uri'] if aad_metadata and 'jwks_uri' in aad_metadata else None
-
             if jwks_uri:
                 response = requests.get(jwks_uri)
                 keys = response.json() if response.ok else None
